requirements.txt
torch==2.2.2
torchvision
numpy
ffmpeg-python
tqdm